fixes and comments

+++ src/pages/Logout.tsx
@@ -1,14 +1,18 @@
-import { useEffect } from "react";
+import { useEffect, useRef } from "react";
 import { toast } from "react-toastify";
 
 const Logout = () => {
+  const hasRun = useRef(false); // useRef doesn’t trigger re-renders
+
   useEffect(() => {
-    console.log("@@@@@@@@");
-    localStorage.clear();
-    toast.success("Logout Successfully");
-    // setTimeout(() => {
-    //   window.location.href = "/";
-    // }, 2000);
+    if (!hasRun.current) {
+      localStorage.clear();
+      toast.success("Logout Successfully");
+      hasRun.current = true;
+      setTimeout(() => {
+        window.location.href = "/";
+      }, 2000);
+    }
   }, []);
 
   return null;
